// src/api-client-react/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Shoes e-commerce API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Product {
  id: number;
  name: string;
  brand: string;
  description: string;
  price: number;
  originalPrice?: number | null;
  imageUrl: string;
  images: string[];
  category: string;
  sizes: number[];
  rating: number;
  reviewCount: number;
  stock: number;
  featured: boolean;
  createdAt: string;
}

export interface CreateProductBody {
  name: string;
  brand: string;
  description: string;
  price: number;
  originalPrice?: number | null;
  imageUrl: string;
  images?: string[];
  category: string;
  sizes: number[];
  stock: number;
  featured: boolean;
}

export interface Category {
  id: number;
  name: string;
  slug: string;
  productCount: number;
}

export interface OrderItem {
  productId: number;
  productName: string;
  productImage: string;
  size: number;
  quantity: number;
  price: number;
}

export type OrderStatus = (typeof OrderStatus)[keyof typeof OrderStatus];

export const OrderStatus = {
  pending: "pending",
  processing: "processing",
  shipped: "shipped",
  delivered: "delivered",
  cancelled: "cancelled",
} as const;

export interface Order {
  id: number;
  customerName: string;
  customerEmail: string;
  customerPhone?: string;
  address: string;
  city: string;
  status: OrderStatus;
  items: OrderItem[];
  subtotal: number;
  delivery: number;
  discount: number;
  total: number;
  createdAt: string;
  updatedAt: string;
}

export type CreateOrderBodyItemsItem = {
  productId: number;
  size: number;
  quantity: number;
};

export interface CreateOrderBody {
  customerName: string;
  customerEmail: string;
  customerPhone?: string;
  address: string;
  city: string;
  items: CreateOrderBodyItemsItem[];
  discountCode?: string | null;
}

export type UpdateOrderStatusBodyStatus =
  (typeof UpdateOrderStatusBodyStatus)[keyof typeof UpdateOrderStatusBodyStatus];

export const UpdateOrderStatusBodyStatus = {
  pending: "pending",
  processing: "processing",
  shipped: "shipped",
  delivered: "delivered",
  cancelled: "cancelled",
} as const;

export interface UpdateOrderStatusBody {
  status: UpdateOrderStatusBodyStatus;
}

export interface DashboardStats {
  totalRevenue: number;
  totalOrders: number;
  totalProducts: number;
  pendingOrders: number;
  revenueChange: number;
  ordersChange: number;
}

export type ListProductsParams = {
  category?: string;
  search?: string;
  featured?: boolean;
  minPrice?: number;
  maxPrice?: number;
};

export type ListOrdersParams = {
  status?: ListOrdersStatus;
};

export type ListOrdersStatus =
  (typeof ListOrdersStatus)[keyof typeof ListOrdersStatus];

export const ListOrdersStatus = {
  pending: "pending",
  processing: "processing",
  shipped: "shipped",
  delivered: "delivered",
  cancelled: "cancelled",
} as const;